backend/requirements.txt
flask
flask-cors
nltk
spacy
pdfminer.six
sentence-transformers
scikit-learn
